vehicles/serializers.py
from rest_framework import serializers
from .models import Brand, Vehicle

class BrandSerializer(serializers.ModelSerializer):
    class Meta:
        model = Brand
        fields = '__all__'

class VehicleSerializer(serializers.ModelSerializer):
    brand_name = serializers.ReadOnlyField(source='brand.name')

    class Meta:
        model = Vehicle
        fields = '__all__'
